New fles

+++ examples/build.sh
@@ -21,7 +21,12 @@ TARGET_DIR="build/$1"
 BASE_PATH=$(pwd)
 
 
-rm -rf $TARGET_DIR
+if [[ $2 != *"keep"* ]]; then
+   echo "Build clean"
+   rm -rf $TARGET_DIR
+fi
+
+
 mkdir -p $TARGET_DIR
 cp -a "../src/" "$TARGET_DIR/tests/"
 cp -a "$SOURCE_DIR/src" $TARGET_DIR
@@ -31,5 +36,23 @@ PATCH_FILE="../../$SOURCE_DIR.patch"
 if [ -e $PATCH_FILE ]; then
    patch -p0 < $PATCH_FILE
 fi
-make -f tests/tools/Makefile test_build test_run test_coverage
 
+if [[ $2 == *"assume_fail"* ]]; then
+   echo "Fail run, assume build ok, running fails"
+   make -f tests/tools/Makefile test_build 
+   for tb in tests/build/bin/*; do
+      set +e
+      ./$tb > /dev/null # Hide the output as its hidious since it fails
+      ret=$?
+      set -e
+      if [[ $ret == 0 ]];then 
+         echo "The test '$tb' did not fail, though it should!"
+         return 1
+      fi;
+   done;
+else
+   echo "Normal run - assume all ok"
+   make -f tests/tools/Makefile test_build test_run test_coverage
+fi
+
+echo "ALL DONE, BYE BYE"
